Add budgeting and subscription management capabilities, improve transaction handling and filtering
- Introduced Budget related entities, DTOs, mappers, and use cases.
- Enabled Subscription management with related entities, DTOs, mappers, and use cases.
- Enhanced TransactionController with new query filters and streamlined endpoint for monthly income and expense.
- Updated entities and services to support comprehensive transaction handling, including budgeted categories and subscriptions.
- Refactored summary and response DTOs for better data representation.
- Enabled scheduling with `@EnableScheduling` in the application.

+++ src/main/java/com/fyrdev/monyia/adapters/driven/jpa/adapter/PocketAdapter.java
@@ -36,7 +36,7 @@ public class PocketAdapter implements IPocketPersistencePort {
     }
 
     @Override
-    public Double getBalance(Long pocketId, Long userId) {
+    public BigDecimal getBalance(Long pocketId, Long userId) {
         return pocketRepository
                 .findBalanceById(pocketId, userId);
     }
